requirements.txt
aiohttp==3.6.3
altair==4.1.0
appdirs==1.4.4
appnope==0.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
async-timeout==3.0.1
attrs==20.2.0
audiotsm==0.1.2
backcall==0.2.0
base58==2.0.1
beautifulsoup4==4.9.3
bleach==3.2.1
blinker==1.4
boto3==1.15.10
botocore==1.18.10
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
coverage==5.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
discord.py==1.5.1
distlib==0.3.1
entrypoints==0.3
enum-compat==0.0.3
filelock==3.0.12
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.9
iniconfig==1.1.1
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
multidict==4.7.6
nbclient==0.5.0
nbconvert==6.0.6
nbformat==5.0.7
nest-asyncio==1.4.1
notebook==6.1.4
numpy==1.18.4
oauthlib==3.1.0
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.2
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pluggy==0.13.1
pprintpp==0.4.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
ptyprocess==0.6.0
py==1.10.0
pyarrow==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==20.7.3
pycountry-convert==0.7.2
pycparser==2.20
pydeck==0.5.0b1
Pygments==2.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.2
pytest-cov==2.11.1
pytest-mock==3.5.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytube==9.6.0
pytube3==9.6.4
pytz==2020.1
pyzmq==19.0.2
repoze.lru==0.7
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.11.0
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
soupsieve==2.2.1
streamlit==0.67.1
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.1
toolz==0.11.1
tornado==6.0.4
traitlets==5.0.4
typing==3.7.4.3
typing-extensions==3.7.4.2
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.9
validators==0.18.1
virtualenv==20.2.1
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
yarl==1.5.1
